refactor and add enforce one tx per signer option
